+++ boxy-core/src/main/java/boxy/core/model/ConsumerGroup.java
@@ -8,7 +8,7 @@ public record ConsumerGroup(
         String name,
         double heartbeatIntervalDefault,
         double heartbeatDeadlineMultiplier,
-        double releaseDeadline,
+        int releaseDeadline,
         int totalWeight,
         int activePartitions,
         int activeWorkers,
